services/order_service/pyproject.toml
[project]
name = "order-service"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.3",
    "django>=6.0.4",
    "djangorestframework>=3.17.1",
    "drf-spectacular>=0.29.0",
    "mysqlclient>=2.2.8",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "requests>=2.33.1",
]
